+++ packages/react-learn-card/src/components/Notification/types.ts
@@ -33,7 +33,7 @@ export const NotificationTypeStyles: {
 } = {
     [NotificationTypeEnum.Currency]: {
         viewButtonStyles: 'border-cyan-400 text-cyan-400',
-        unclaimedButtonStyles: 'text-white bg-cyan-400 border-cyan-400',
+        unclaimedButtonStyles: 'text-white bg-cyan-400 border-cyan-400 shadow-bottom',
         claimedButtonStyles: 'text-cyan-400 bg-cyan-50 border-cyan-50',
         textStyles: 'text-cyan-400 capitalize',
         iconCircleStyles: 'bg-cyan-400',
@@ -42,7 +42,7 @@ export const NotificationTypeStyles: {
     },
     [NotificationTypeEnum.SocialBadges]: {
         viewButtonStyles: 'border-cyan-400 text-cyan-400',
-        unclaimedButtonStyles: 'text-white bg-cyan-400 border-cyan-400',
+        unclaimedButtonStyles: 'text-white bg-cyan-400 border-cyan-400 shadow-bottom',
         claimedButtonStyles: 'text-cyan-400 bg-cyan-50 border-cyan-50',
         textStyles: 'text-cyan-400 capitalize',
         iconCircleStyles: 'bg-cyan-400',
@@ -51,7 +51,7 @@ export const NotificationTypeStyles: {
     },
     [NotificationTypeEnum.ID]: {
         viewButtonStyles: 'border-yellow-400 text-yellow-400',
-        unclaimedButtonStyles: 'text-white bg-yellow-400 border-yellow-400',
+        unclaimedButtonStyles: 'text-white bg-yellow-400 border-yellow-400 shadow-bottom',
         claimedButtonStyles: 'text-yellow-400 bg-yellow-50 border-yellow-50',
         textStyles: 'text-yellow-400 uppercase',
         iconCircleStyles: 'bg-yellow-400',
@@ -60,7 +60,7 @@ export const NotificationTypeStyles: {
     },
     [NotificationTypeEnum.Achievement]: {
         viewButtonStyles: 'border-spice-600 text-spice-400',
-        unclaimedButtonStyles: 'text-white bg-spice-400 border-spice-400',
+        unclaimedButtonStyles: 'text-white bg-spice-400 border-spice-400 shadow-bottom',
         claimedButtonStyles: 'text-spice-400 bg-spice-50 border-spice-50',
         textStyles: 'text-spice-400 capitalize',
         iconCircleStyles: 'bg-spice-400',
@@ -69,7 +69,7 @@ export const NotificationTypeStyles: {
     },
     [NotificationTypeEnum.Skill]: {
         viewButtonStyles: 'border-indigo-400 text-indigo-400',
-        unclaimedButtonStyles: 'text-white bg-indigo-400 border-indigo-400',
+        unclaimedButtonStyles: 'text-white bg-indigo-400 border-indigo-400 shadow-bottom',
         claimedButtonStyles: 'text-indigo-400 bg-indigo-50 border-indigo-50',
         textStyles: 'text-indigo-400 capitalize',
         iconCircleStyles: 'bg-indigo-400',
@@ -78,7 +78,7 @@ export const NotificationTypeStyles: {
     },
     [NotificationTypeEnum.Job]: {
         viewButtonStyles: 'border-rose-400 text-rose-400',
-        unclaimedButtonStyles: 'text-white bg-rose-400 border-rose-400',
+        unclaimedButtonStyles: 'text-white bg-rose-400 border-rose-400 shadow-bottom',
         claimedButtonStyles: 'text-rose-400 bg-rose-50 border-rose-50',
         textStyles: 'text-rose-400 capitalize',
         iconCircleStyles: 'bg-rose-400',
@@ -87,7 +87,7 @@ export const NotificationTypeStyles: {
     },
     [NotificationTypeEnum.Learning]: {
         viewButtonStyles: 'border-emerald-500 text-emerald-500',
-        unclaimedButtonStyles: 'text-white bg-emerald-500 border-emerald-500',
+        unclaimedButtonStyles: 'text-white bg-emerald-500 border-emerald-500 shadow-bottom',
         claimedButtonStyles: 'text-emerald-500 bg-emerald-50 border-emerald-50',
         textStyles: 'text-emerald-500 capitalize',
         iconCircleStyles: 'bg-emerald-500',