+++ packages/react-learn-card/src/components/NotificationUserAcceptedBoostCard/types.ts
@@ -1,0 +1,13 @@
+import { UserNotificationTypeEnum } from '../../constants/notifications';
+
+export type NotificationUserAcceptedBoostCardProps = {
+    className?: string;
+    thumbImage?: string;
+    customThumbComponent?: React.ReactNode;
+    notificationType?: UserNotificationTypeEnum.AcceptedBoost;
+    issueDate?: string;
+    handleCancelClick?: () => void;
+    loadingState?: boolean;
+    title?: string;
+    isArchived?: boolean;
+};