+++ examples/snap-chapi-example/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @learncard/snap-chapi-example
 
+## 1.0.55
+
+### Patch Changes
+
+-   Updated dependencies [[`5cc4fce`](https://github.com/learningeconomy/LearnCard/commit/5cc4fce1fcdfaedf6d50e5e7e99df37ff06ae393)]:
+    -   @learncard/react@2.5.12
+
+## 1.0.54
+
+### Patch Changes
+
+-   Updated dependencies [[`fdbce48`](https://github.com/learningeconomy/LearnCard/commit/fdbce48c2c0a467c1ae480f6dcc4ae377de51aab)]:
+    -   @learncard/react@2.5.11
+
 ## 1.0.53
 
 ### Patch Changes