+++ packages/react-learn-card/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @welibraryos/react-learn-card
 
+## 2.5.12
+
+### Patch Changes
+
+-   [#270](https://github.com/learningeconomy/LearnCard/pull/270) [`5cc4fce`](https://github.com/learningeconomy/LearnCard/commit/5cc4fce1fcdfaedf6d50e5e7e99df37ff06ae393) Thanks [@goblincore](https://github.com/goblincore)! - fix: Export notification card components properly
+
+## 2.5.11
+
+### Patch Changes
+
+-   [#268](https://github.com/learningeconomy/LearnCard/pull/268) [`fdbce48`](https://github.com/learningeconomy/LearnCard/commit/fdbce48c2c0a467c1ae480f6dcc4ae377de51aab) Thanks [@goblincore](https://github.com/goblincore)! - Add updated LCA notification card components
+
 ## 2.5.10
 
 ### Patch Changes